raspberry-pi/requirements.txt
# SmartAmbient Raspberry Pi Hub
# Python application for IoT hub functionality

paho-mqtt>=2.0.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
uuid>=1.30
